+++ common/time/Date.hpp
@@ -12,18 +12,18 @@ namespace common::io
     {
     public:
         Date();
-        Date(int year, int month, int day);
-        Date(int year, int month, int day, int hours, int minutes, int seconds);
-        explicit Date(long long timestamp);
+        Date(int32_t year, int32_t month, int32_t day);
+        Date(int32_t year, int32_t month, int32_t day, int32_t hours, int32_t minutes, int32_t seconds);
+        explicit Date(int64_t timestamp);
         ~Date() = default;
         [[nodiscard]] auto clone() const -> Date;
         [[nodiscard]] auto equals(const Date& other) const -> bool;
         [[nodiscard]] auto after(const Date& other) const -> bool;
         [[nodiscard]] auto before(const Date& other) const -> bool;
-        [[nodiscard]] auto getTime() const -> long long;
-        [[nodiscard]] auto getYear() const -> int;
-        [[nodiscard]] auto getMonth() const -> int;
-        [[nodiscard]] auto getDay() const -> int;
+        [[nodiscard]] auto getTime() const -> int64_t;
+        [[nodiscard]] auto getYear() const -> int32_t;
+        [[nodiscard]] auto getMonth() const -> int32_t;
+        [[nodiscard]] auto getDay() const -> int32_t;
         [[nodiscard]] auto toString() const -> std::string;
         [[nodiscard]] auto hashCode() const -> size_t;
     private:
@@ -33,7 +33,7 @@ namespace common::io
 
     inline Date::Date(): time_point_(std::chrono::system_clock::now()) {}
 
-    inline Date::Date(const int year, const int month, const int day)
+    inline Date::Date(const int32_t year, const int32_t month, const int32_t day)
     {
         if (month < 1 || month > 12 || day < 1 || day > 31)
         {
@@ -46,7 +46,7 @@ namespace common::io
         time_point_ = std::chrono::system_clock::from_time_t(std::mktime(&tm));
     }
 
-    inline Date::Date(const int year, const int month, const int day, const int hours, const int minutes, const int seconds)
+    inline Date::Date(const int32_t year, const int32_t month, const int32_t day, const int32_t hours, const int32_t minutes, const int32_t seconds)
     {
         if (month < 1 || month > 12 || day < 1 || day > 31 || hours < 0 || hours > 23 || minutes < 0 || minutes > 59 || seconds < 0 || seconds > 59)
         {
@@ -62,7 +62,7 @@ namespace common::io
         time_point_ = std::chrono::system_clock::from_time_t(std::mktime(&tm));
     }
 
-    inline Date::Date(const long long timestamp)
+    inline Date::Date(const int64_t timestamp)
     {
         time_point_ = std::chrono::system_clock::time_point(std::chrono::milliseconds(timestamp));
     }
@@ -87,22 +87,22 @@ namespace common::io
         return time_point_ < other.time_point_;
     }
 
-    inline auto Date::getTime() const -> long long
+    inline auto Date::getTime() const -> int64_t
     {
         return std::chrono::duration_cast<std::chrono::milliseconds>(time_point_.time_since_epoch()).count();
     }
 
-    inline auto Date::getYear() const -> int
+    inline auto Date::getYear() const -> int32_t
     {
         return toTm().tm_year + 1900;
     }
 
-    inline auto Date::getMonth() const -> int
+    inline auto Date::getMonth() const -> int32_t
     {
         return toTm().tm_mon + 1;
     }
 
-    inline auto Date::getDay() const -> int
+    inline auto Date::getDay() const -> int32_t
     {
         return toTm().tm_mday;
     }
@@ -117,7 +117,7 @@ namespace common::io
 
     inline auto Date::hashCode() const -> size_t
     {
-        return std::hash<long long>()(getTime());
+        return std::hash<int64_t>()(getTime());
     }
 
     inline auto Date::toTm() const -> std::tm
